Added skip delete to images save script

+++ docker/save.sh
@@ -15,6 +15,7 @@
 
 # set default values for pulling images
 IMAGE_TAG="latest"
+SKIP_DELETE="false"
 
 # override defaults from arguments
 while [ "$1" != "" ]; do
@@ -26,6 +27,10 @@ while [ "$1" != "" ]; do
               IMAGE_TAG="latest"
             fi
             ;;
+        -s | --skip-delete )
+            SKIP_DELETE="true"
+            echo "Images will not be deleted"
+            ;;            
         * )
             echo "Invalid parameter: "$1
             exit 1
@@ -41,11 +46,13 @@ echo "Using Images Tag: ${IMAGE_TAG}"
 images="contivvpp/ksr:${IMAGE_TAG} contivvpp/cni:${IMAGE_TAG} contivvpp/stn:${IMAGE_TAG} contivvpp/vswitch:${IMAGE_TAG}"
 echo $images
 if [ -f ../vagrant/images.tar ]; then
-        rm ../vagrant/images.tar
+  rm ../vagrant/images.tar
 fi
 
 docker save $images -o ../vagrant/images.tar
 
-for img in $images; do
-        docker rmi $img
-done
+if [ "${SKIP_DELETE}" != "true" ]; then
+  for img in $images; do
+    docker rmi $img
+  done
+fi
